Re-Worked Movement and Food Pickup
+updated to 2019.1f1
-need to change the food scale equation

+++ Consumo/Assets/Scripts/GameManager.cs
@@ -6,16 +6,16 @@ using UnityEngine.UI;
 public class GameManager : MonoBehaviour
 {
     public int currentLives;
-    public int globalWeight;
+    public float globalWeight;
     public Text weightText;
 
-    public void AddWeight(int weightToAdd)
+    public void AddWeight(float weightToAdd)
     {
         globalWeight += weightToAdd;
         weightText.text = "Global Weight: " + globalWeight;
     }
 
-    public void RemoveWeight(int weightToRemove)
+    public void RemoveWeight(float weightToRemove)
     {
         globalWeight -= weightToRemove;
         weightText.text = "Global Weight: " + globalWeight;
@@ -28,21 +28,4 @@ public class GameManager : MonoBehaviour
         }
         
     }
-    //replace spawnpoint with array
-    //for(int i = 0; i < players.length; i++)
-    //player[i].transform.position = playerSpawnPoint[i].position
-
-    //private TankManager GetRoundWinner()
-    //{
-    //    // Go through all the sumos...
-    //    for (int i = 0; i < m_Tanks.Length; i++)
-    //    {
-    //        // ... and if one of them is active, it is the winner so return it.
-    //        if (m_Tanks[i].m_Instance.activeSelf)
-    //            return m_Tanks[i];
-    //    }
-
-    //    // If none of the tanks are active it is a draw so return null.
-    //    return null;
-    //}
 }
